+++ src/Runner.php
@@ -10,7 +10,7 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 
 class Runner {
 
-    const VERSION = '0.4.0';
+    const VERSION = '0.4.1';
     const ROBOCLASS = 'RoboFile';
     const ROBOFILE = 'RoboFile.php';